services/flask/requirements.txt
flake8==3.7.8
Flask==1.1.1
python-dateutil==2.8.1
pytz==2019.3
